proper reassignments and a slight refactor

+++ game/scene_managers/rendering_scripts/standard_rendering.cpp
@@ -30,7 +30,7 @@ namespace rendering_scripts {
 		const auto& physics = cosmos.systems_temporary.get<physics_system>();
 		const auto& controlled_entity = cosmos[step.viewed_character];
 		const auto& interp = step.session.systems_audiovisual.get<interpolation_system>();
-		const float global_time_seconds = step.get_interpolated_total_time_passed_in_seconds();
+		const float global_time_seconds = static_cast<float>(step.get_interpolated_total_time_passed_in_seconds());
 
 		auto all_visible = dynamic_tree.determine_visible_entities_from_camera(camera);
 		const auto visible_from_physics = physics.query_camera(camera).entities;
